switch to using class attributes instead of string column names.  this simplifies the code internally (e.g., is_fk is not needed since we can just call the appropriate method) and externally (since errors of using an invalid column are now obvious).
I tried to standardize on "colname" in the code for where we still use strings occasionally, and "col" for Column objects.  Early on though the object is actually an InstrumentedAttribute, which allows specifying relations as well, but calling those "columns" is a little confusing.  I need to go through the code and change those variable and method names to attr.