Added insert_after() on the FieldSet object.  Fixed display bug.  Tweaked docs.  Tests included.
The display bug was just showing off the wrong object when raising the TypeError.

The reason for adding insert_after(), is that in some cases, you want to add a field related to the previous field
and not the next field.  For example, you have a `password` field.  If most of the time, you just want a single field
for the password to appear, then you let it that way.  If at some point, you want to have a second confirmation password
shown (while in a user's profile edit mode, or something alike), you will add it *after* the first password box, and not
necessarily before the next field, which might not always be the same, according to some other business logic.