#84 Make sure we can call .configure() multiple times, with consistent and predictable results.
Before this patch, the outcome of calling configure() twice was unpredictable, thus
no old code should be broken.

Removed some documentation to ModelRenderer for params it didn't have.
NOTE: there is duplication with documentation in docs/forms.txt from ModelRenderer.

It is now possible to call a fs.configure(options=[fs.field1.required()]) and it
will simply replace the `field1` in the _render_fields.  This will work only if
you do not specify `include` and `exclude`.

Using this feature, you can create FieldSet configured with the full fields.  Then
go back and tweak certain fields.   NOTE: This is a little bit less useful since we
have set(option=blah) on the Fields, that modify fields in-place.

Also, it will save focus=True/False and readonly=True/False between calls to
configure().  You'd have to specify it explicitly to change the values.  The default
for each remains the same as before.  The same is true for global_validator on
AbstractFieldSet and FieldSet.