Implement an optional tiny deserialization cache.  Documented.
Tests pass, though no new tests were added to make sure this doesn't break in the future.

The reason to implement this is that the deserialize() function gets called twice in
a normal controller.  Once with validate() and a second time with sync().

For custom Renderers that have complex deserialization (fetching to a database, creating
temporary objects, appending to a DB relation), you don't want it executed twice.

Using this simple deserialize_once (imported from formalchemy.fields) decorator
solves this problem.